backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.1
google-cloud-speech==2.24.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.5
numpy==1.26.4
orjson==3.9.14
packaging==23.2
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic-core==2.16.2
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pytest==8.0.1
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
tomli==2.0.1
typing-extensions==4.9.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
